Update action bar to use diff viewing preference
Use the diff viewing preference instead of only having a view in browser option
Remove the status stripe from the properties card as it is shown on the commit card
