+++ app/src/main/res/menu/change_details_menu.xml
@@ -26,7 +26,7 @@
 
     <item
         android:id="@+id/menu_details_browser"
-        android:icon="?attr/browserIcon"
+        android:icon="?attr/viewIcon"
         android:showAsAction="ifRoom"
         android:title="@string/view_in_browser"/>
 </menu>
